+++ build.js
@@ -36,7 +36,7 @@ var dirBundle = function ( dir, cb ) {
 
 var features           = dirBundle( './src/features/' );
 var plugins            = dirBundle( './src/plugins/' );
-var path_resolver      = fs.readFileSync( './src/path.resolver.js' );
+var path_resolver      = fs.readFileSync( './src/otpl.resolver.js' );
 var sample_output_file = '"'
 	+ fs.readFileSync( './src/output.js.sample' )
 		.toString()
@@ -50,7 +50,7 @@ var bundle = fs.readFileSync( './src/otpl.base.js', 'utf8' )
 			   .replace( "{OTPL::SAMPLE_OUTPUT_FILE}", sample_output_file )
 			   .replace( "{OTPL::FEATURES}", indent( features ) )
 			   .replace( "{OTPL::PLUGINS}", indent( plugins ) )
-			   .replace( "{OTPL::PATH_RESOLVER}", indent( path_resolver ) );
+			   .replace( "{OTPL::RESOLVER}", indent( path_resolver ) );
 
 if ( !fs.existsSync( './build' ) ) {
 	fs.mkdirSync( './build' );