+++ php/src/OTplData.php
@@ -1,6 +1,11 @@
 <?php
+/**
+ * Copyright (c) Emile Silas Sare <emile.silas@gmail.com>
+ *
+ * This file is part of Otpl.
+ */
 
-	final class OTplData {
+final class OTplData {
 		private $data = null;
 		private $context = null;