+++ src/features/import/otpl.import.js
@@ -1,15 +1,29 @@
+/*
+ * Copyright (c) Emile Silas Sare <emile.silas@gmail.com>
+ *
+ * This file is part of Otpl.
+ */
+
 //Ex: @import(url,data) --> OTplUtils.importExec(url,data)
 OTplUtils.addReplacer( {
-	reg : /@import\([\s]*?(?:'|")(.*?)(?:'|")(?:[\s]*,[\s]*(.+?)[\s]*)?[\s]*\)/g,
+	reg : /@import\([\s]*?(['"]?(.*?)['"]?)(?:[\s]*,[\s]*(.+?)[\s]*)?[\s]*\)/g,
 	val : function () {
 		var scope = this,
 			_root = scope.getSrcDir() || scope.getRootDir(),
-			url   = arguments[ 1 ],
-			data  = arguments[ 2 ];
+			isExpression = (arguments[ 1 ] === arguments[ 2 ]),
+			url   = arguments[ 2 ],
+			data  = arguments[ 3 ];
 
-		url = OPathResolver.resolve( _root, url );
 
-		if ( /\.otpl$/.test( url ) ) {
+		if( isExpression ){
+			var expression = arguments[ 1 ];
+
+			return "OTplUtils.importCustom( '" + escape(_root) + "', " + expression + "," + data + ")";
+		}
+
+		url = OTplResolver.resolve( _root, url );
+
+		if ( _endsWith( url , '.otpl' ) ) {
 			var child = (new OTpl()).parse( url );
 
 			scope.addDependancies( child.getDependancies() );